+++ app/src/androidTest/java/ch/epfl/sdp/utils/MockMapApi.java
@@ -1,7 +1,5 @@
 package ch.epfl.sdp.utils;
 
-import android.util.Log;
-
 import androidx.test.core.app.ApplicationProvider;
 
 import org.junit.Test;
@@ -9,12 +7,11 @@ import org.junit.Test;
 import java.util.ArrayList;
 import java.util.Collection;
 
-import ch.epfl.sdp.SocialTests.MockFriendsFetcher;
 import ch.epfl.sdp.dependencies.MyApplication;
 import ch.epfl.sdp.entity.Player;
 import ch.epfl.sdp.game.Game;
 import ch.epfl.sdp.geometry.GeoPoint;
-import ch.epfl.sdp.item.Market;
+import ch.epfl.sdp.market.Market;
 import ch.epfl.sdp.map.Displayable;
 import ch.epfl.sdp.map.MapApi;
 import ch.epfl.sdp.map.Renderer;
