+++ app/src/main/java/ch/epfl/sdp/dependencies/AppContainer.java
@@ -6,11 +6,9 @@ import ch.epfl.sdp.database.firebase.api.CommonDatabaseAPI;
 import ch.epfl.sdp.database.firebase.api.CommonFirestoreDatabaseAPI;
 import ch.epfl.sdp.database.firebase.api.ServerDatabaseAPI;
 import ch.epfl.sdp.database.firebase.api.ServerFirestoreDatabaseAPI;
-import ch.epfl.sdp.item.Market;
+import ch.epfl.sdp.market.Market;
 import ch.epfl.sdp.login.AuthenticationAPI;
 import ch.epfl.sdp.login.FirebaseAuthenticationAPI;
-import ch.epfl.sdp.map.GoogleMapApi;
-import ch.epfl.sdp.map.MapApi;
 import ch.epfl.sdp.map.MapsActivity;
 import ch.epfl.sdp.market.MarketActivity;
 import ch.epfl.sdp.social.Conversation.FireStoreToSQLiteAdapter;
