+++ app/src/main/java/ch/epfl/sdp/game/Server.java
@@ -13,7 +13,7 @@ import ch.epfl.sdp.geometry.RectangleArea;
 import ch.epfl.sdp.geometry.UnboundedArea;
 import ch.epfl.sdp.item.Healthpack;
 import ch.epfl.sdp.item.ItemBox;
-import ch.epfl.sdp.item.Market;
+import ch.epfl.sdp.market.Market;
 
 /**
  * Takes care of all actions that a server should perform (generating enemies, updating enemies etc.).
