+++ app/src/main/java/ch/epfl/sdp/market/MarketActivity.java
@@ -12,14 +12,12 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import java.util.HashMap;
-import java.util.Map;
 
 import ch.epfl.sdp.R;
 import ch.epfl.sdp.dependencies.MyApplication;
 import ch.epfl.sdp.entity.PlayerManager;
 import ch.epfl.sdp.item.Healthpack;
 import ch.epfl.sdp.item.Item;
-import ch.epfl.sdp.item.Market;
 import ch.epfl.sdp.item.Scan;
 import ch.epfl.sdp.item.Shield;
 import ch.epfl.sdp.item.Shrinker;
